// frontend/src/api/generated/models/TripRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TripRequest = {
    arrival_location_id?: string;
    arrival_time?: string | null;
    comment?: string | null;
    departure_location_id?: string;
    departure_time?: string | null;
    series_id?: string | null;
    status?: string;
    transport_type_id?: number;
};